Adding navigation compose and implementing it for each module
