+++ features/profile/src/main/java/com/mmk/profile/presentation/ProfileFragment.kt
@@ -3,29 +3,10 @@ package com.mmk.profile.presentation
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontStyle
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.fragment.app.Fragment
-import com.mmk.common.ui.theme.MyApplicationTheme
-import com.mmk.common.ui.theme.OrangeDark
-import com.mmk.common.ui.theme.OrangeLight
-import com.mmk.profile.R
 
 class ProfileFragment : Fragment() {
 
@@ -36,52 +17,5 @@ class ProfileFragment : Fragment() {
     ) = ComposeView(requireContext()).apply {
         setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
         setContent { ProfileScreen() }
-    }
-
-    @Composable
-    fun ProfileScreen() {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(24.dp),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.ic_person),
-                contentDescription = null,
-                modifier = Modifier.size(100.dp)
-            )
-            Text(
-                text = stringResource(id = R.string.profile_name),
-                style = MaterialTheme.typography.titleLarge.copy(fontSize = 20.sp),
-                modifier = Modifier.padding(top = 16.dp),
-            )
-            Box(
-                contentAlignment = Alignment.Center,
-                modifier = Modifier
-                    .padding(top = 8.dp)
-                    .border(1.dp, OrangeDark, RoundedCornerShape(8.dp))
-                    .background(
-                        color = OrangeLight,
-                        shape = RoundedCornerShape(8.dp)
-                    )
-                    .height(68.dp)
-                    .fillMaxWidth()
-            ) {
-                Text(
-                    text = stringResource(id = R.string.profile_bio),
-                    fontStyle = FontStyle.Italic,
-                    style = MaterialTheme.typography.titleMedium.copy(fontSize = 16.sp),
-                )
-            }
-        }
-    }
-
-    @Preview(showBackground = true)
-    @Composable
-    private fun ProfileScreenPreview() {
-        MyApplicationTheme {
-            ProfileScreen()
-        }
     }
 }
