+++ features/quotes/src/main/java/com/mmk/quotes/domain/usecase/addquote/AddNewQuoteImpl.kt
@@ -6,6 +6,7 @@ import com.mmk.quotes.domain.repository.QuotesRepository
 
 class AddNewQuoteImpl(private val quotesRepository: QuotesRepository) : AddNewQuote {
     override suspend fun invoke(quote: Quote): Result<Unit> {
+
         return quotesRepository.addNewQuote(quote)
     }
 }
